requirements.txt
selenium
beautifulsoup4
lxml
python-dotenv
